Enforce Django 1.6.5 dependency.
It has been pointed ouy (h/t @stevejalim) that since South has been
rolled into 1.7, having a dependency on South would cause problems. I
have therefore pinned this version of the app to 1.6.5, as I know that
works. As soon as we have upgraded to Django 1.7, I'll remove the
dependency on South, and enforce a dependency on 1.7. No, it won't be
backwards compatible. Sorry about that.