Change the way in which profiling is cancelled.
Iterating through the signal responses was an inelegant solution, and it
put the onus on the client developer to remember to return True from
their signal receiver function, which tripped me up a couple of times.

I've updated the ProfilerRecord model to add an explicit `cancel()`
method which can be called by the signal receiver function to cancel
the profile.