Change the way that global exclusions are applied.
Instead of a single IGNORE_STAFF setting, there is now setting called
REQUEST_PROFILER_GLOBAL_EXCLUDE_FUNC, which is a function that is
called on each request, and that can be used to exclude requests from
profiling. By default this is a lambda that return True, so all
requests are profiled.