Refactor internals & drop Py37,38 (#31)
This PR reworks the middleware and changes the names of the
ProfilingRecord model so that they match the middleware methods.
(set_request -> process_request etc.)

There is a breaking change in the way in which the middleware works, in
that it now follows the middleware structure, process_request,
process_view, process_response, and it no longer passes the request
object to the process_response method. It used to do this so that the
request was guaranteed to have the session / user attributes set, even
if the middleware ran after the session / auth middleware (as
process_response happens on the way out, so all of the
process_request methods will have run by then).

This is convenient, but a hack. In the new world, if you put this
middleware before the django middleware you will not record the session
/ user info, as it won't be set. This is by design. The new
recommendation is that you put this middleware after Django middleware,
as these aren't really optional, and there's very little you can do to
optimise this code. Profile the bits you can edit, ignore the bits you
can't. If Django itself is slowing you down, then that's a whole other
problem.